+++ libraries/transformer/src/test/java/androidx/media3/transformer/CompositionPlayerTest.java
@@ -19,6 +19,7 @@ import static androidx.media3.common.Player.STATE_BUFFERING;
 import static androidx.media3.common.Player.STATE_ENDED;
 import static androidx.media3.common.Player.STATE_IDLE;
 import static androidx.media3.common.Player.STATE_READY;
+import static androidx.media3.test.utils.TestUtil.createByteCountingAudioProcessor;
 import static androidx.media3.test.utils.TestUtil.getCommandsAsList;
 import static androidx.media3.test.utils.robolectric.RobolectricUtil.runMainLooperUntil;
 import static androidx.media3.test.utils.robolectric.TestPlayerRunHelper.advance;
@@ -1112,6 +1113,33 @@ public class CompositionPlayerTest {
         new Composition.Builder(new EditedMediaItemSequence.Builder(item).build()).build();
 
     assertThrows(IllegalArgumentException.class, () -> player.setComposition(composition));
+  }
+
+  @Test
+  public void setSpeed_onSecondarySequence_outputsExpectedNumberOfSamples() throws Exception {
+    AtomicInteger bytes = new AtomicInteger();
+    AudioProcessor processor = createByteCountingAudioProcessor(bytes);
+    CompositionPlayer player = createTestCompositionPlayer();
+    EditedMediaItem item =
+        new EditedMediaItem.Builder(MediaItem.fromUri(ASSET_URI_PREFIX + FILE_AUDIO_RAW))
+            .setDurationUs(1_000_000L)
+            .build();
+    EditedMediaItem speedAdjustedItem =
+        item.buildUpon()
+            .setSpeed(TestSpeedProvider.createWithStartTimes(new long[] {0L}, new float[] {0.5f}))
+            .build();
+    EditedMediaItemSequence primarySequence = new EditedMediaItemSequence.Builder(item).build();
+    EditedMediaItemSequence secondarySequence =
+        new EditedMediaItemSequence.Builder(speedAdjustedItem).build();
+    player.setComposition(
+        new Composition.Builder(primarySequence, secondarySequence)
+            .setEffects(new Effects(ImmutableList.of(processor), ImmutableList.of()))
+            .build());
+    player.prepare();
+
+    play(player).untilState(STATE_ENDED);
+
+    assertThat(bytes.get() / 2).isEqualTo(88200);
   }
 
   private static Composition buildComposition() {